// ReformaAgraria/ClientApp/app/models/gen/persilSummaryViewModel.ts
export interface PersilSummaryViewModel { 
    totalTransmigrasi?: number;
    totalProna?: number;
    totalKonflik?: number;
    totalTerlantar?: number;
    totalPelepasan?: number;
    totalAdat?: number;
    totalDesa?: number;
    totalTanamanRakyat?: number;
    totalKemasyarakatan?: number;
    totalKemitraan?: number;
}